Building Database
There's a lot to work on with building the database. I've added a solid
schema and data generation script, but it will take a few days to clean
up the word lists. I can't even imagine what I'll have to come up with
to make clues, but I won't focus on that for now. Worst case scenario is
that I come up with a system to auto-generate puzzles, and I come up
with the clues myself for a while.

The project is a little messed up at the moment. I've got a DIY repo
pattern that can handle grabbing words from the DB in a convenient way,
and I wrote it in TypeScript. The backend server code is in JavaScript.
I failed to consider this. May be best to recreate the backend
scaffolding using TS, since it doesn't have any logic yet. The repos and
their interfaces should belong to the backend, I don't think they have
anything to do with pure Crossword puzzle logic which is in the
`Crossword` folder.

TODO: Recreate backend w/ TS scaffolding
TODO: Get `DictionaryRepo` class and interface out of `Crossword` folder