Crossword Generation Success
Was able to speed up the backtracking alg
significantly by performing a bit of preproccesing
on the words in the database.

The database now contains a table with all
possible characters which can follow a given
prefix character sequence and a given number
of remaining spaces to make a word.

The backtracking alg uses this to quickly get
a constrained character set, and uses some more
tricks to constrain the character set further
given the current row and column it is building.

After some tests, the generation routine
exhibited an incredibly large range of total
run times, ranging from up to an hour (with no
success, the process was killed before it
finished) down to almost instant success. It was
also very common for the process to instantly
finish. This led to the last trick.

A time limit was added, which if exceeded,
causes the entire grid to be reset and the
process to start again. This was found to
greatly reduce the average time to find
a solution, taking no longer than 30 seconds
for 99% of runs. (Not an official stat, just
an observation.)

It could even generate a 9x9 grid in under 30
seconds. For different size grids, the time limit
can be tuned. For a 6x6 grid, a 5 second time
limit is more than enough.